requirements.txt
requests
discord.py (Lisez "LisezMoi.md" pour plus d'informations)
discord.py[voice] (Lisez "LisezMoi.md" pour plus d'informations)